pyproject.toml
[project]
name = "end-to-end-data-pipeline"
version = "2.0.0"
description = "Production-grade end-to-end data pipeline: Airflow + Kafka + Spark + Postgres + MinIO + MLflow + Great Expectations + Snowflake + Prometheus + Grafana, with a Python FastAPI backend ported 1:1 from a .NET 8 reference. A learnwithparam.com workshop."
authors = [{ name = "Param Harrison", email = "param@learnwithparam.com" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "structlog>=24.1.0",

    # Databases
    "asyncpg>=0.29.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "aiomysql>=0.2.0",
    "pymysql>=1.1.0",

    # Object storage
    "minio>=7.2.0",
    "boto3>=1.34.0",

    # Streaming
    "aiokafka>=0.10.0",
    "confluent-kafka>=2.3.0",

    # ML / observability
    "mlflow>=2.10.0",
    "prometheus-client>=0.20.0",

    # Data quality
    "great-expectations>=0.18.0",

    # Data
    "pandas>=2.2.0",
    "pyarrow>=15.0.0",
]

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 120
target-version = "py311"
